Deduplicate meta and progress databases.

Remove duplicated entries in the meta and progress databases in order to lighten storage. These databases are located in the _targets/meta/meta and _targets/meta/progress files, and the _targets data store lives in the current working directory. No essential data is removed, so this is simply a form of garbage collection.

tar_deduplicate(meta = TRUE, progress = TRUE)

Arguments

meta
Logical, whether to deduplicate the meta database file at _targets/meta/meta.

progress
Logical, whether to deduplicate the progress database file at _targets/meta/progress.

Value

Nothing.

Examples

if (FALSE) {
tar_dir({
tar_script(tar_pipeline(tar_target(x, 1 + 1)))
tar_make()
tar_make()
tar_deduplicate() # Compare the file _targets/meta/meta before and after.
})
}