tar_deps {targets}	R Documentation

Code dependencies

Description

List the dependencies of a function or expression.

Usage

tar_deps(expr)

Arguments

expr
A quoted R expression or function.

Details

targets detects the dependencies of commands using static code analysis. Use tar_deps() to run the code analysis and see the dependencies for yourself.

Value

Character vector of the dependencies of a function or expression.

See Also

Other inspect: tar_deps_raw(), tar_manifest(), tar_network(), tar_outdated(), tar_sitrep(), tar_validate()

Examples

tar_deps(x <- y + z)
tar_deps({
  x <- 1
  x + a
})
tar_deps(function(a = b) map_dfr(data, ~do_row(.x)))

[Package targets version 1.0.0 Index]