tar_deps_raw {targets}	R Documentation

Code dependencies (raw version)

Description

Same as tar_deps() except expr must already be an unquoted function or expression object.

Usage

tar_deps_raw(expr)

Arguments

expr
An R expression object or function.

Value

Character vector of the dependencies of a function or expression.

See Also

Other inspect: tar_deps(), tar_manifest(), tar_network(), tar_outdated(), tar_sitrep(), tar_validate()

Examples

tar_deps_raw(quote(x <- y + z))
tar_deps_raw(
  quote({
    x <- 1
    x + a
  })
)
tar_deps_raw(function(a = b) map_dfr(data, ~do_row(.x)))

[Package targets version 1.0.0 Index]