tar_dir {targets}	R Documentation

Execute code in a temporary directory.

Description

Runs code inside a new tempfile() directory in order to avoid writing to the user's file space. Used in examples and tests in order to comply with CRAN policies.

Usage

tar_dir(code)

Arguments

code
User-defined code.

Value

Return value of the user-defined code.

See Also

Other utilities to extend targets: tar_assert, tar_condition, tar_language, tar_test()

Examples

tar_dir(file.create("only_exists_in_tar_dir"))
file.exists("only_exists_in_tar_dir")

[Package targets version 1.0.0 Index]