tar_edit {targets}	R Documentation

Open the target script file for editing.

Description

Open the target script file for editing. Requires the usethis package.

Usage

tar_edit(script = targets::tar_config_get("script"))

Arguments

script
Character of length 1, path to the target script file. Defaults to tar_config_get("script"), which in turn defaults to _targets.R. When you set this argument, the value of tar_config_get("script") is temporarily changed for the current function call. See tar_script(), tar_config_get(), and tar_config_set() for details about the target script file and how to set it persistently for a project.

Details

The target script file is an R code file that defines the pipeline. The default path is _targets.R, but the default for the current project can be configured with tar_config_set().

See Also

Other scripts: tar_github_actions(), tar_helper_raw(), tar_helper(), tar_renv(), tar_script()

[Package targets version 1.0.0 Index]