For developers only: get the environment of the current target.

For developers only: get the environment where a target runs its command. Inherits from tar_option_get("envir").

tar_envir(default = parent.frame())

Arguments

default
Environment, value to return if tar_envir() is called on its own outside a targets pipeline. Having a default lets users run things without tar_make(), which helps peel back layers of code and troubleshoot bugs.

Value

If called from a running target, tar_envir() returns the environment where the target runs its command. If called outside a pipeline, the return value is whatever the user supplies to default (which defaults to parent.frame()).

Details

Users should not call tar_envir() directly because accidental modifications to parent.env(tar_envir()) could break the pipeline. tar_envir() only exists in order to support third-party interface packages such as tarchetypes.

Examples

tar_envir()

#> <environment: 0x7fc757706030>

tar_envir(default = new.env(parent = emptyenv()))

#> <environment: 0x7fc756c75bf8>

tar_target(name, tar_envir(default = parent.frame()))

#> <stem target> 
#>   name: name 
#>   command:
#>     tar_envir(default = parent.frame()) 
#>   format: rds 
#>   iteration method: vector 
#>   error mode: stop 
#>   memory mode: persistent 
#>   storage mode: main 
#>   retrieval mode: main 
#>   deploy to: worker 
#>   resources:
#>     list() 
#>   cue:
#>     mode: thorough
#>     command: TRUE
#>     depend: TRUE
#>     format: TRUE
#>     iteration: TRUE
#>     file: TRUE 
#>   packages:
#>     targets
#>     stats
#>     graphics
#>     grDevices
#>     utils
#>     datasets
#>     methods
#>     base 
#>   library:
#>     NULL