tar_exist_process {targets}	R Documentation

Check if process metadata exists.

Description

Check if the process metadata file _targets/meta/process exists for the current project.

Usage

tar_exist_process(store = targets::tar_config_get("store"))

Arguments

store
Character of length 1, path to the targets data store. Defaults to tar_config_get("store"), which in turn defaults to _targets/. When you set this argument, the value of tar_config_get("store") is temporarily changed for the current function call. See tar_config_get() and tar_config_set() for details about how to set the data store path persistently for a project.

Details

To learn more about data storage in targets, visit https://books.ropensci.org/targets/data.html.

Value

Logical of length 1, whether the current project's metadata exists.

See Also

Other existence: tar_exist_meta(), tar_exist_objects(), tar_exist_progress(), tar_exist_script()

Examples

tar_exist_process()

[Package targets version 1.0.0 Index]