tar_helper_raw {targets}	R Documentation

Write a helper R script (raw version).

Description

Write a helper R script for a targets pipeline. Could be supporting functions or the target script file (default: _targets.R) itself.

Usage

tar_helper_raw(path = NULL, code = NULL)

Arguments

path
Character of length 1, path to write (or overwrite) code. If the parent directory does not exist, tar_helper_raw() creates it.

code
Expression object. tar_helper_raw() deparses and writes this code to a file at path, overwriting it if the file already exists.

Details

tar_helper_raw() is a specialized version of tar_script() with flexible paths and tidy evaluation. It is like tar_helper() except that code is an "evaluated" argument rather than a quoted one.

Value

NULL (invisibly)

See Also

Other scripts: tar_edit(), tar_github_actions(), tar_helper(), tar_renv(), tar_script()

Examples

path <- tempfile()
tar_helper_raw(path, quote(x <- 1))
writeLines(readLines(path))

[Package targets version 1.0.0 Index]