Run a pipeline of targets.

Run the pipeline you defined in _targets.R. tar_make() runs the correct targets in the correct order and stores the return values in _targets/objects/.

tar_make(
  names = NULL,
  reporter = "verbose",
  garbage_collection = FALSE,
  callr_function = callr::r,
  callr_arguments = list()
)

Arguments

names
Names of the targets to build or check. Set to NULL to check/build all the targets (default). Otherwise, you can supply symbols, a character vector, or tidyselect helpers like starts_with().

reporter
Character of length 1, name of the reporter to user. Controls how messages are printed as targets run in the pipeline. Choices:

"verbose": print one message for each target that runs (default).

"silent": print nothing.

"timestamp": print a time-stamped message for each target that runs.

"summary": print a running total of the number of each targets in each status category (queued, running, skipped, build, cancelled, or errored).

garbage_collection
Logical, whether to run base::gc() between targets. The pipeline will run slower but consume less memory.

callr_function
A function from callr to start a fresh clean R process to do the work. Set to NULL to run in the current session instead of an external process (but restart your R session just before you do in order to clear debris out of the global environment). callr_function needs to be NULL for interactive debugging, e.g. tar_option_set(debug = "your_target"). However, callr_function should not be NULL for serious reproducible work.

callr_arguments
A list of arguments to callr_function.

Value

NULL except if callr_function = callr::r_bg(), in which case a handle to the callr background process is returned. Either way, the value is invisibly returned.

Examples

tar_dir({
tar_script({
  tar_option_set()
  tar_pipeline(tar_target(x, 1 + 1))
})
tar_make()
tar_script({
  tar_option_set()
  tar_pipeline(
    tar_target(y1, 1 + 1),
    tar_target(y2, 1 + 1),
    tar_target(z, y1 + y2)
  )
})
tar_make(starts_with("y")) # Only builds y1 and y2.
})

#> ● run target x
#> ● run target y1
#> ● run target y2
#>