Return the vertices and edges of a pipeline dependency graph.

Analyze the pipeline defined in _targets.R and return the vertices and edges of the directed acyclic graph of dependency relationships.

tar_network(
  targets_only = FALSE,
  reporter = "silent",
  callr_function = callr::r,
  callr_arguments = list()
)

Arguments

targets_only
Logical, whether to restrict the output to just targets (FALSE) or to also include imported global functions and objects.

reporter
Character of length 1, name of the reporter to user. Controls how messages are printed as targets are checked. Choices:

"silent": print nothing.

"forecast": print running totals of the checked and outdated targets found so far.

callr_function
A function from callr to start a fresh clean R process to do the work. Set to NULL to run in the current session instead of an external process (but restart your R session just before you do in order to clear debris out of the global environment). callr_function needs to be NULL for interactive debugging, e.g. tar_option_set(debug = "your_target"). However, callr_function should not be NULL for serious reproducible work.

callr_arguments
A list of arguments to callr_function.

Value

A list with two data frames: vertices and edges. The vertices data frame has one row per target with fields to denote the type of the target or object (stem, branch, map, cross, function, or object) and the target's status (up to date, outdated, running, cancelled, or errored). The edges data frame has one row for every edge and columns to and from to mark the starting and terminating vertices.

Examples

if (FALSE) {
tar_dir({
tar_script({
  tar_option_set()
  tar_pipeline(
    tar_target(y1, 1 + 1),
    tar_target(y2, 1 + 1),
    tar_target(z, y1 + y2)
  )
})
tar_network(targets_only = TRUE)
})
}