tar_objects {targets}	R Documentation

List saved targets

Description

List targets currently saved to _targets/objects/ or the cloud. Does not include local files with tar_target(..., format = "file", repository = "local").

Usage

tar_objects(
  names = NULL,
  cloud = TRUE,
  store = targets::tar_config_get("store")
)

Arguments

names
Optional tidyselect selector such as any_of() or starts_with() to return a tactical subset of target names. If NULL, all names are selected.

cloud
Logical of length 1, whether to include cloud targets in the output (e.g. tar_target(..., repository = "aws")).

store
Character of length 1, path to the targets data store. Defaults to tar_config_get("store"), which in turn defaults to _targets/. When you set this argument, the value of tar_config_get("store") is temporarily changed for the current function call. See tar_config_get() and tar_config_set() for details about how to set the data store path persistently for a project.

Value

Character vector of targets saved to _targets/objects/.

See Also

Other data: tar_load_everything(), tar_load_raw(), tar_load(), tar_meta(), tar_pid(), tar_process(), tar_read_raw(), tar_read()

Examples

if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) { # for CRAN
tar_dir({ # tar_dir() runs code from a temp dir for CRAN.
tar_script({
  list(tar_target(x, "value"))
}, ask = FALSE)
tar_make()
tar_objects()
tar_objects(starts_with("x")) # see also any_of()
})
}

[Package targets version 1.0.0 Index]