Get a target option.

Get a target option. These options include default arguments to tar_target() such as packages, storage format, iteration type, and cue. Needs to be called before any calls to tar_target() in order to take effect.

tar_option_get(option)

Arguments

option
Character of length 1, name of an option to get. Must be one of the argument names of tar_option_set().

Value

Value of a target option.

Details

This function goes well with tar_target_raw() when it comes to defining external interfaces on top of the targets package to create pipelines.

Examples

tar_option_get("format") # default format before we set anything

#> [1] "rds"

tar_target(x, 1)$settings$format

#> [1] "rds"

tar_option_set(format = "fst_tbl") # new default format
tar_option_get("format")

#> [1] "fst_tbl"

tar_target(x, 1)$settings$format

#> [1] "fst_tbl"

tar_option_reset() # reset the format
tar_target(x, 1)$settings$format

#> [1] "rds"