tar_path_store {targets}	R Documentation

Current data store path

Description

Identify the file path to the data store of the pipeline currently running.

Usage

tar_path_store()

Value

Character, file path to the data store of the pipeline currently running. If called outside of the pipeline currently running, tar_path_store() returns tar_config_get("store").

See Also

Other utilities: tar_active(), tar_call(), tar_cancel(), tar_definition(), tar_envir(), tar_group(), tar_name(), tar_path_script_support(), tar_path_script(), tar_path_target(), tar_path(), tar_seed(), tar_source(), tar_store()

Examples

tar_path_store()
if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) { # for CRAN
tar_dir({ # tar_dir() runs code from a temp dir for CRAN.
tar_script(tar_target(x, tar_path_store()), ask = FALSE)
store <- tempfile()
tar_make(store = store)
tar_read(x, store = store)
})
}

[Package targets version 1.0.0 Index]