Declare a pipeline of targets.

A pipeline is similar to a Makefile or a drake plan. It is a collection of targets, or skippable steps, in an automated reproducible workflow.

tar_pipeline(...)

Arguments

...
Targets or lists of targets defined with tar_target().

Value

A pipeline object. The _targets.R file of a project must end with a call tar_pipeline(). Otherwise, users do not work pipeline objects directly.

Examples

file <- tar_target(data_file, "data.csv", format = "file")
object <- tar_target(data_object, read.csv(data_file))
analysis <- tar_target(analysis, analyze(data_object))
pipeline <- tar_pipeline(file, object, analysis)
print(pipeline)

#> <pipeline with 3 targets>

# Equivalent:
target_list <- list(
  tar_target(data_file, "data.csv", format = "file"),
  tar_target(data_object, read.csv(data_file)),
  tar_target(analysis, analyze(data_object))
)
pipeline <- tar_pipeline(target_list)