tar_progress_branches {targets}	R Documentation

Tabulate the progress of dynamic branches.

Description

Read a project's target progress data for the most recent run of the pipeline and display the tabulated status of dynamic branches. Only the most recent record is shown.

Usage

tar_progress_branches(
  names = NULL,
  fields = NULL,
  store = targets::tar_config_get("store")
)

Arguments

names
Optional, names of the targets. If supplied, tar_progress() only returns progress information on these targets. You can supply symbols or tidyselect helpers like starts_with().

fields
Optional, names of progress data columns to read. Set to NULL to read all fields.

store
Character of length 1, path to the targets data store. Defaults to tar_config_get("store"), which in turn defaults to _targets/. When you set this argument, the value of tar_config_get("store") is temporarily changed for the current function call. See tar_config_get() and tar_config_set() for details about how to set the data store path persistently for a project.

Value

A data frame with one row per target per progress status and the following columns.

name: name of the pattern.

progress: progress status: "started", "built", "cancelled", or "errored".

branches: number of branches in the progress category.

total: total number of branches planned for the whole pattern. Values within the same pattern should all be equal.

See Also

Other progress: tar_built(), tar_canceled(), tar_errored(), tar_poll(), tar_progress_summary(), tar_progress(), tar_skipped(), tar_started(), tar_watch_server(), tar_watch_ui(), tar_watch()

Examples

if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) { # for CRAN
tar_dir({ # tar_dir() runs code from a temp dir for CRAN.
tar_script({
  list(
    tar_target(x, seq_len(2)),
    tar_target(y, x, pattern = map(x)),
    tar_target(z, stopifnot(y < 1.5), pattern = map(y))
  )
}, ask = FALSE)
try(tar_make())
tar_progress_branches()
})
}

[Package targets version 1.0.0 Index]