tar_progress_summary {targets}	R Documentation

Summarize target progress.

Description

Summarize the progress of a run of the pipeline.

Usage

tar_progress_summary(
  fields = c("skipped", "started", "built", "errored", "canceled", "since"),
  store = targets::tar_config_get("store")
)

Arguments

fields
Optional, names of progress data columns to read. Set to NULL to read all fields.

store
Character of length 1, path to the targets data store. Defaults to tar_config_get("store"), which in turn defaults to _targets/. When you set this argument, the value of tar_config_get("store") is temporarily changed for the current function call. See tar_config_get() and tar_config_set() for details about how to set the data store path persistently for a project.

Value

A data frame with one row and the following optional columns that can be selected with fields. (time is omitted by default.)

started: number of targets that started and did not (yet) finish.

built: number of targets that completed without error or cancellation.

errored: number of targets that threw an error.

canceled: number of canceled targets (see tar_cancel()).

since: how long ago progress last changed (Sys.time() - time).

time: the time when the progress last changed (modification timestamp of the _targets/meta/progress file).

See Also

Other progress: tar_built(), tar_canceled(), tar_errored(), tar_poll(), tar_progress_branches(), tar_progress(), tar_skipped(), tar_started(), tar_watch_server(), tar_watch_ui(), tar_watch()

Examples

if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) { # for CRAN
tar_dir({ # tar_dir() runs code from a temp dir for CRAN.
tar_script({
  list(
    tar_target(x, seq_len(2)),
    tar_target(y, x, pattern = map(x)),
    tar_target(z, stopifnot(y < 1.5), pattern = map(y), error = "continue")
  )
}, ask = FALSE)
try(tar_make())
tar_progress_summary()
})
}

[Package targets version 1.0.0 Index]