tar_resources_gcp {targets}	R Documentation

Target resources: Google Cloud Platform (GCP) Google Cloud Storage (GCS)

Description

Create the gcp argument of tar_resources() to specify optional settings for Google Cloud Storage for targets with tar_target(..., repository = "gcp"). See the format argument of tar_target() for details.

Usage

tar_resources_gcp(
  bucket = targets::tar_option_get("resources")$gcp$bucket,
  prefix = targets::tar_option_get("resources")$gcp$prefix,
  predefined_acl = targets::tar_option_get("resources")$gcp$predefined_acl,
  verbose = targets::tar_option_get("resources")$gcp$verbose
)

Arguments

bucket
Character of length 1, name of an existing bucket to upload and download the return values of the affected targets during the pipeline.

prefix
Character of length 1, "directory path" in the bucket where the target return values are stored. Defaults to targets::tar_path_objects_dir_cloud().

predefined_acl
Character of length 1, user access to the object. See ?googleCloudStorageR::gcs_upload for possible values. Defaults to "private".

verbose
Logical of length 1, whether to print extra messages like progress bars during uploads and downloads. Defaults to FALSE.

Details

See the cloud storage section of https://books.ropensci.org/targets/data.html for details for instructions.

Value

Object of class "tar_resources_gcp", to be supplied to the gcp argument of tar_resources().

Resources

Functions tar_target() and tar_option_set() each takes an optional resources argument to supply non-default settings of various optional backends for data storage and high-performance computing. The tar_resources() function is a helper to supply those settings in the correct manner.

In targets version 0.12.2 and above, resources are inherited one-by-one in nested fashion from tar_option_get("resources"). For example, suppose you set tar_option_set(resources = tar_resources(aws = my_aws)), where my_aws equals tar_resources_aws(bucket = "x", prefix = "y"). Then, tar_target(data, get_data() will have bucket "x" and prefix "y". In addition, if new_resources equals tar_resources(aws = tar_resources_aws(bucket = "z"))), then tar_target(data, get_data(), resources = new_resources) will use the new bucket "z", but it will still use the prefix "y" supplied through tar_option_set(). (In targets 0.12.1 and below, options like prefix do not carry over from tar_option_set() if you supply non-default resources to tar_target().)

See Also

Other resources: tar_resources_aws(), tar_resources_clustermq(), tar_resources_crew(), tar_resources_feather(), tar_resources_fst(), tar_resources_future(), tar_resources_parquet(), tar_resources_qs(), tar_resources_url(), tar_resources()

Examples

# Somewhere in you target script file (usually _targets.R):
tar_target(
  name,
  command(),
  format = "qs",
  repository = "gcp",
  resources = tar_resources(
    gcp = tar_resources_gcp(bucket = "yourbucketname"),
    qs = tar_resources_qs(preset = "fast")
  )
)

[Package targets version 1.0.0 Index]