tar_resources_url {targets}	R Documentation

Target resources: URL storage formats

Description

Create the url argument of tar_resources() to specify optional settings for URL storage formats. See the format argument of tar_target() for details.

Usage

tar_resources_url(
  handle = targets::tar_option_get("resources")$url$handle,
  seconds_interval = 1,
  seconds_timeout = 10
)

Arguments

handle
Object returned by curl::new_handle or NULL. Defaults to NULL.

seconds_interval
Nonnegative numeric of length 1, number of seconds to wait between individual retries while attempting to connect to the URL.

seconds_timeout
Nonnegative numeric of length 1, number of seconds to wait before timing out while trying to connect to the URL.

Value

Object of class "tar_resources_url", to be supplied to the url argument of tar_resources().

Resources

Functions tar_target() and tar_option_set() each takes an optional resources argument to supply non-default settings of various optional backends for data storage and high-performance computing. The tar_resources() function is a helper to supply those settings in the correct manner.

In targets version 0.12.2 and above, resources are inherited one-by-one in nested fashion from tar_option_get("resources"). For example, suppose you set tar_option_set(resources = tar_resources(aws = my_aws)), where my_aws equals tar_resources_aws(bucket = "x", prefix = "y"). Then, tar_target(data, get_data() will have bucket "x" and prefix "y". In addition, if new_resources equals tar_resources(aws = tar_resources_aws(bucket = "z"))), then tar_target(data, get_data(), resources = new_resources) will use the new bucket "z", but it will still use the prefix "y" supplied through tar_option_set(). (In targets 0.12.1 and below, options like prefix do not carry over from tar_option_set() if you supply non-default resources to tar_target().)

See Also

Other resources: tar_resources_aws(), tar_resources_clustermq(), tar_resources_crew(), tar_resources_feather(), tar_resources_fst(), tar_resources_future(), tar_resources_gcp(), tar_resources_parquet(), tar_resources_qs(), tar_resources()

Examples

if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) { # for CRAN
# Somewhere in you target script file (usually _targets.R):
tar_target(
  name,
  command(),
  format = "url",
  resources = tar_resources(
    url = tar_resources_url(handle = curl::new_handle())
  )
)
}

[Package targets version 1.0.0 Index]