tar_test {targets}	R Documentation

Test code in a temporary directory.

Description

Runs a test_that() unit test inside a temporary directory to avoid writing to the user's file space. This helps ensure compliance with CRAN policies. Also isolates tar_option_set() options and environment variables specific to targets and skips the test on Solaris. Useful for writing tests for targetopia packages (extensions to targets tailored to specific use cases).

Usage

tar_test(label, code)

Arguments

label
Character of length 1, label for the test.

code
User-defined code for the test.

Value

NULL (invisibly).

See Also

Other utilities to extend targets: tar_assert, tar_condition, tar_dir(), tar_language

Examples

tar_test("example test", {
  testing_variable_cafecfcb <- "only defined inside tar_test()"
  file.create("only_exists_in_tar_test")
})
exists("testing_variable_cafecfcb")
file.exists("only_exists_in_tar_test")

[Package targets version 1.0.0 Index]