tar_workspaces {targets}	R Documentation

List saved target workspaces.

Description

List target workspaces currently saved to _targets/workspaces/. See tar_workspace() for more information.

Usage

tar_workspaces(names = NULL, store = targets::tar_config_get("store"))

Arguments

names
Optional tidyselect selector to return a tactical subset of workspace names. If NULL, all names are selected.

store
Character of length 1, path to the targets data store. Defaults to tar_config_get("store"), which in turn defaults to _targets/. When you set this argument, the value of tar_config_get("store") is temporarily changed for the current function call. See tar_config_get() and tar_config_set() for details about how to set the data store path persistently for a project.

Value

Character vector of available workspaces to load with tar_workspace().

See Also

Other debug: tar_load_globals(), tar_traceback(), tar_workspace()

Examples

if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) { # for CRAN
tar_dir({ # tar_dir() runs code from a temp dir for CRAN.
tar_script({
  tar_option_set(workspace_on_error = TRUE)
  list(
    tar_target(x, "value"),
    tar_target(y, x)
  )
}, ask = FALSE)
tar_make()
tar_workspaces()
tar_workspaces(contains("x"))
})
}

[Package targets version 1.0.0 Index]