use_targets_rmd {targets}	R Documentation

Use targets with Target Markdown.

Description

Create an example Target Markdown report to get started with targets.

Usage

use_targets_rmd(path = "_targets.Rmd", open = interactive())

Arguments

path
Character of length 1, output path of the Target Markdown report relative to the current active project.

open
Logical, whether to open the file for editing in the RStudio IDE.

Value

NULL (invisibly).

See Also

Other help: tar_reprex(), targets-package, use_targets()

Examples

if (identical(Sys.getenv("TAR_INTERACTIVE_EXAMPLES"), "true")) {
tar_dir({ # tar_dir() runs code from a temp dir for CRAN.
use_targets(open = FALSE)
})
}

[Package targets version 1.0.0 Index]