tar_branch_names {targets}	R Documentation

Branch names

Description

Get the branch names of a dynamic branching target using numeric indexes.

Usage

tar_branch_names(name, index, store = targets::tar_config_get("store"))

Arguments

name
Symbol, name of the dynamic branching target (pattern).

index
Integer vector of branch indexes.

store
Character of length 1, path to the targets data store. Defaults to tar_config_get("store"), which in turn defaults to _targets/. When you set this argument, the value of tar_config_get("store") is temporarily changed for the current function call. See tar_config_get() and tar_config_set() for details about how to set the data store path persistently for a project.

Value

A character vector of branch names.

See Also

Other branching: tar_branch_index(), tar_branch_names_raw(), tar_branches(), tar_pattern()

Examples

if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) { # for CRAN
tar_dir({ # tar_dir() runs code from a temp dir for CRAN.
tar_script({
  list(
    tar_target(x, seq_len(4)),
    tar_target(y, 2 * x, pattern = map(x)),
    tar_target(z, y, pattern = map(y))
  )
}, ask = FALSE)
tar_make()
tar_branch_names(z, c(2, 3))
})
}

[Package targets version 1.0.0 Index]