tar_load {targets}	R Documentation

Load the values of targets.

Description

Load the return values of targets into the current environment (or the environment of your choosing). For a typical target, the return value lives in a file in _targets/objects/. For dynamic files (i.e. format = "file") the paths loaded in place of the values. tar_load_everything() is shorthand for tar_load(everything()) to load all targets.

Usage

tar_load(
  names,
  branches = NULL,
  meta = tar_meta(targets_only = TRUE, store = store),
  strict = TRUE,
  silent = FALSE,
  envir = parent.frame(),
  store = targets::tar_config_get("store")
)

Arguments

names
Names of the targets to load. You may supply tidyselect helpers like any_of() and starts_with(). Names are selected from the metadata in _targets/meta, which may include errored targets.

branches
Integer of indices of the branches to load for any targets that are patterns.

meta
Data frame of metadata from tar_meta(). tar_read() with the default arguments can be inefficient for large pipelines because all the metadata is stored in a single file. However, if you call tar_meta() beforehand and supply it to the meta argument, then successive calls to tar_read() may run much faster.

strict
Logical of length 1, whether to error out if one of the selected targets is in the metadata but cannot be loaded. Set to FALSE to just load the targets in the metadata that can be loaded and skip the others.

silent
Logical of length 1. Only relevant when strict is FALSE. If silent is FALSE and strict is FALSE, then a message will be printed if a target is in the metadata but cannot be loaded. However, load failures will not stop other targets from being loaded.

envir
Environment to put the loaded targets.

store
Character of length 1, path to the targets data store. Defaults to tar_config_get("store"), which in turn defaults to _targets/. When you set this argument, the value of tar_config_get("store") is temporarily changed for the current function call. See tar_config_get() and tar_config_set() for details about how to set the data store path persistently for a project.

Value

Nothing.

Limited scope

tar_read() and tar_load() are only for exploratory analysis and literate programming, and tar_read_raw() and tar_load_raw() are only for exploratory analysis. targets automatically loads the correct dependencies into memory when the pipeline is running, so invoking these functions from inside a target is rarely advisable.

See Also

Other data: tar_load_everything(), tar_load_raw(), tar_meta(), tar_objects(), tar_pid(), tar_process(), tar_read_raw(), tar_read()

Examples

if (identical(Sys.getenv("TAR_EXAMPLES"), "true")) { # for CRAN
tar_dir({ # tar_dir() runs code from a temp dir for CRAN.
tar_script({
  list(
    tar_target(y1, 1 + 1),
    tar_target(y2, 1 + 1),
    tar_target(z, y1 + y2)
  )
}, ask = FALSE)
tar_make()
ls() # Does not have "y1", "y2", or "z".
tar_load(starts_with("y"))
ls() # Has "y1" and "y2" but not "z".
tar_load(any_of("z"))
ls() # Has "y1", "y2", and "z".
})
}

[Package targets version 1.0.0 Index]