Show the cue-by-cue status of each target.

For each target, report which cues are activated. Except for the never cue, the target will rerun in tar_make() if any cue is activated. The target is suppressed if the never cue is TRUE. See tar_cue() for details.

tar_sitrep(
  names = NULL,
  fields = NULL,
  reporter = "silent",
  callr_function = callr::r,
  callr_arguments = list(spinner = identical(reporter, "silent"))
)

Arguments

names
Optional, names of the targets. If supplied, tar_sitrep() only returns metadata on these targets. You can supply symbols, a character vector, or tidyselect helpers like starts_with().

fields
Optional, names of columns/fields to select. If supplied, tar_sitrep() only returns the selected metadata columns. You can supply symbols, a character vector, or tidyselect helpers like starts_with(). The name column is always included first no matter what you select. Choices:

name: name of the target or global object.

record: Whether the record cue is activated: TRUE if the target is not in the metadata (tar_meta()), or if the target errored during the last tar_make(), or if the class of the target changed.

always: Whether mode in tar_cue() is "always". If TRUE, tar_make() always runs the target.

never: Whether mode in tar_cue() is "never". If TRUE, tar_make() will only run if the record cue activates.

command: Whether the target's command changed since last time. Always TRUE if the record cue is activated. Otherwise, always FALSE if the command cue is suppressed.

depend: Whether the data/output of at least one of the target's dependencies changed since last time. Dependencies are targets, functions, and global objects directly upstream. Call tar_outdated(targets_only = FALSE) or tar_visnetwork(targets_only = FALSE) to see exactly which dependencies are outdated. Always NA if the record cue is activated. Otherwise, always FALSE if the depend cue is suppressed.

format: Whether the storage format of the target is different from last time. Always NA if the record cue is activated. Otherwise, always FALSE if the format cue is suppressed.

iteration: Whether the iteration mode of the target is different from last time. Always NA if the record cue is activated. Otherwise, always FALSE if the iteration cue is suppressed.

file: Whether the file(s) with the target's return value are missing or different from last time. Always NA if the record cue is activated. Otherwise, always FALSE if the file cue is suppressed.

reporter
Character of length 1, name of the reporter to user. Controls how messages are printed as targets are checked. Choices:

"silent": print nothing.

"forecast": print running totals of the checked and outdated targets found so far.

callr_function
A function from callr to start a fresh clean R process to do the work. Set to NULL to run in the current session instead of an external process (but restart your R session just before you do in order to clear debris out of the global environment). callr_function needs to be NULL for interactive debugging, e.g. tar_option_set(debug = "your_target"). However, callr_function should not be NULL for serious reproducible work.

callr_arguments
A list of arguments to callr_function.

Value

A data frame with one row per target/object and one column per cue. Each element is a logical to indicate whether the cue is activated for the target. See the field argument in this help file for details.

Details

Caveats:

tar_cue() allows you to change/suppress cues, so the return value will depend on the settings you supply to tar_cue().

If a pattern tries to branches over a target that does not exist in storage, then the branches are omitted from the output.

tar_sitrep() is myopic. It only considers what happens to the immediate target and its immediate upstream dependencies, and it makes no attempt to propagate invalidation downstream.

Examples

if (identical(Sys.getenv("TARGETS_LONG_EXAMPLES"), "true")) {
tar_dir({
tar_script(
  tar_pipeline(
    tar_target(x, seq_len(2)),
    tar_target(y, 2 * x, pattern = map(x))
  )
)
tar_make()
tar_sitrep()
tar_meta(starts_with("y_"))
})
}