Validate a pipeline of targets.

Inspect the pipeline for issues and throw an error or warning if a problem is detected.

tar_validate(callr_function = callr::r, callr_arguments = list())

Arguments

callr_function
A function from callr to start a fresh clean R process to do the work. Set to NULL to run in the current session instead of an external process (but restart your R session just before you do in order to clear debris out of the global environment). callr_function needs to be NULL for interactive debugging, e.g. tar_options(debug = "your_target"). However, callr_function should not be NULL for serious reproducible work.

callr_arguments
A list of arguments to callr_function.

Value

Nothing.

Examples

if (FALSE) {
tar_dir({
tar_script(tar_pipeline(tar_target(x, 1 + 1)))
tar_validate()
})
}