Visualize an abridged fast dependency graph.

Analyze the pipeline defined in _targets.R and visualize the directed acyclic graph of targets and imported global functions and objects.

tar_visnetwork(
  targets_only = FALSE,
  allow = NULL,
  exclude = NULL,
  outdated = TRUE,
  label = NULL,
  reporter = "silent",
  callr_function = callr::r,
  callr_arguments = list(spinner = reporter == "silent")
)

Arguments

targets_only
Logical, whether to restrict the output to just targets (FALSE) or to also include imported global functions and objects.

allow
Optional, define the set of allowable vertices in the graph. Set to NULL to allow all vertices in the pipeline and environment (default). Otherwise, you can supply symbols, a character vector, or tidyselect helpers like starts_with().

exclude
Optional, define the set of exclude vertices from the graph. Set to NULL to exclude no vertices. Otherwise, you can supply symbols, a character vector, or tidyselect helpers like starts_with().

outdated
Logical, whether to show colors to distinguish outdated targets from up-to-date targets. (Global functions and objects still show these colors.) Looking for outdated targets takes a lot of time for large pipelines with lots of branches, and setting outdated to FALSE is a nice way to speed up the graph if you only want to see dependency relationships and build progress.

label
Character vector of one or more aesthetics to add to the vertex labels. Can contain "time" to show total runtime, "size" to show total storage size, or "branches" to show the number of branches in each pattern. You can choose multiple aesthetics at once, e.g. label = c("time", "branches"). All are disabled by default because they clutter the graph.

reporter
Character of length 1, name of the reporter to user. Controls how messages are printed as targets are checked. Choices:

"silent": print nothing.

"forecast": print running totals of the checked and outdated targets found so far.

callr_function
A function from callr to start a fresh clean R process to do the work. Set to NULL to run in the current session instead of an external process (but restart your R session just before you do in order to clear debris out of the global environment). callr_function needs to be NULL for interactive debugging, e.g. tar_option_set(debug = "your_target"). However, callr_function should not be NULL for serious reproducible work.

callr_arguments
A list of arguments to callr_function.

Value

A visNetwork HTML widget object.

Examples

if (FALSE) {
tar_dir({
tar_script({
  tar_option_set()
  tar_pipeline(
    tar_target(y1, 1 + 1),
    tar_target(y2, 1 + 1),
    tar_target(z, y1 + y2)
  )
})
tar_visnetwork()
tar_visnetwork(allow = starts_with("y"))
})
}