Shiny app to watch the dependency graph.

Calls tar_visnetwork() every few seconds.

tar_watch(
  seconds = 5,
  seconds_min = 1,
  seconds_max = 100,
  seconds_step = 1,
  targets_only = FALSE,
  outdated = TRUE,
  label = NULL,
  level_separation = 150,
  height = "700px"
)

Arguments

seconds
Numeric of length 1, default number of seconds between refreshes of the graph. Can be changed in the app controls.

seconds_min
Numeric of length 1, lower bound of seconds in the app controls.

seconds_max
Numeric of length 1, upper bound of seconds in the app controls.

seconds_step
Numeric of length 1, step size of seconds in the app controls.

targets_only
Logical, whether to restrict the output to just targets (FALSE) or to also include imported global functions and objects.

outdated
Logical, whether to show colors to distinguish outdated targets from up-to-date targets. (Global functions and objects still show these colors.) Looking for outdated targets takes a lot of time for large pipelines with lots of branches, and setting outdated to FALSE is a nice way to speed up the graph if you only want to see dependency relationships and build progress.

label
Label for the module.

level_separation
Numeric of length 1, levelSeparation argument of visNetwork::visHierarchicalLayout(). Controls the distance between hierarchical levels. Consider changing the value if the aspect ratio of the graph is far from 1. If level_separation is NULL, the levelSeparation argument of visHierarchicalLayout() defaults to 150.

height
Character of length 1, height of the visNetwork widget.

Value

A Shiny app.

Details

The controls of the app are in the left panel. The seconds control is the number of seconds between refreshes of the graph, and the other settings match the arguments of tar_visnetwork().

Examples

if (FALSE) { # Only run interactively.
tar_dir({
tar_script({
  sleep_run <- function(...) {
    Sys.sleep(10)
  }
  tar_pipeline(
    tar_target(settings, sleep_run()),
    tar_target(data1, sleep_run(settings)),
    tar_target(data2, sleep_run(settings)),
    tar_target(data3, sleep_run(settings)),
    tar_target(model1, sleep_run(data1)),
    tar_target(model2, sleep_run(data2)),
    tar_target(model3, sleep_run(data3)),
    tar_target(figure1, sleep_run(model1)),
    tar_target(figure2, sleep_run(model2)),
    tar_target(figure3, sleep_run(model3)),
    tar_target(conclusions, sleep_run(c(figure1, figure2, figure3)))
  )
})
px <- tar_make(callr_function = callr::r_bg)
tar_watch(seconds = 10, outdated = FALSE, targets_only = TRUE)
})
}