Shiny module UI for tar_watch()

Use tar_watch_ui() and tar_watch_server() to include tar_watch() as a Shiny module in an app.

tar_watch_ui(
  id,
  label = "tar_watch_label",
  seconds = 5,
  seconds_min = 1,
  seconds_max = 60,
  seconds_step = 1,
  targets_only = FALSE,
  outdated = TRUE,
  label_tar_visnetwork = NULL,
  level_separation = 150,
  height = "700px"
)

Arguments

id
An ID string that corresponds with the ID used to call the module's UI function.

label
Label for the module.

seconds
Numeric of length 1, default number of seconds between refreshes of the graph. Can be changed in the app controls.

seconds_min
Numeric of length 1, lower bound of seconds in the app controls.

seconds_max
Numeric of length 1, upper bound of seconds in the app controls.

seconds_step
Numeric of length 1, step size of seconds in the app controls.

targets_only
Logical, whether to restrict the output to just targets (FALSE) or to also include imported global functions and objects.

outdated
Logical, whether to show colors to distinguish outdated targets from up-to-date targets. (Global functions and objects still show these colors.) Looking for outdated targets takes a lot of time for large pipelines with lots of branches, and setting outdated to FALSE is a nice way to speed up the graph if you only want to see dependency relationships and build progress.

label_tar_visnetwork
Character vector, label argument to tar_visnetwork().

level_separation
Numeric of length 1, levelSeparation argument of visNetwork::visHierarchicalLayout(). Controls the distance between hierarchical levels. Consider changing the value if the aspect ratio of the graph is far from 1. If level_separation is NULL, the levelSeparation argument of visHierarchicalLayout() defaults to 150.

height
Character of length 1, height of the visNetwork widget.

Value

A Shiny module UI.

Examples

str(tar_watch_ui("my_id"))

#> List of 3
#>  $ name    : chr "div"
#>  $ attribs :List of 1
#>   ..$ class: chr "row"
#>  $ children:List of 2
#>   ..$ :List of 3
#>   .. ..$ name    : chr "div"
#>   .. ..$ attribs :List of 1
#>   .. .. ..$ class: chr "col-sm-4"
#>   .. ..$ children:List of 1
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ :List of 3
#>   .. .. .. .. ..$ name    : chr "head"
#>   .. .. .. .. ..$ attribs : Named list()
#>   .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. ..$ name    : chr "style"
#>   .. .. .. .. .. .. ..$ attribs : Named list()
#>   .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. ..$ : 'html' chr ".direct-chat-contacts {\n                 -webkit-transform: translate(100%, 0);\n                 -ms-transfor"| __truncated__
#>   .. .. .. .. .. .. .. .. ..- attr(*, "html")= logi TRUE
#>   .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. ..- attr(*, "htmltools.singleton")= logi TRUE
#>   .. .. .. ..$ :List of 3
#>   .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. ..$ class: chr "col-sm-12"
#>   .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. ..$ class: chr "card card-primary"
#>   .. .. .. .. .. .. ..$ children:List of 3
#>   .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. .. .. ..$ class: chr "card-header"
#>   .. .. .. .. .. .. .. .. ..$ children:List of 2
#>   .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "h3"
#>   .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "card-title"
#>   .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Control"
#>   .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "card-tools"
#>   .. .. .. .. .. .. .. .. .. .. ..$ children:List of 6
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "button"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type            : chr "button"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class           : chr "btn btn-tool"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-card-widget: chr "collapse"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "i"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "fa fa-minus"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "html_dependencies")=List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 10
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name      : chr "font-awesome"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ version   : chr "5.13.0"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ src       :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ file: chr "www/shared/fontawesome"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ meta      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ script    : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ stylesheet: chr [1:2] "css/all.min.css" "css/v4-shims.min.css"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ head      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attachment: NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ package   : chr "shiny"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ all_files : logi TRUE
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "html_dependency"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "browsable_html")= logi TRUE
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ : NULL
#>   .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. .. .. ..$ attribs :List of 2
#>   .. .. .. .. .. .. .. .. .. ..$ class  : chr "card-body"
#>   .. .. .. .. .. .. .. .. .. ..$ inputID: chr "my_id-control"
#>   .. .. .. .. .. .. .. .. ..$ children:List of 6
#>   .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-group shiny-input-container"
#>   .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "control-label"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ for  : chr "my_id-seconds"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "seconds"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 13
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class                  : chr "js-range-slider"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id                     : chr "my_id-seconds"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-min               : chr "1"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-max               : chr "60"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-from              : chr "5"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-step              : chr "1"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-grid              : chr "true"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-grid-num          : num 9.83
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-grid-snap         : chr "false"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-prettify-separator: chr ","
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-prettify-enabled  : chr "true"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-keyboard          : chr "true"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-data-type         : chr "number"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. ..- attr(*, "html_dependencies")=List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 10
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ name      : chr "ionrangeslider"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ version   : chr "2.1.6"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ src       :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ href: chr "shared/ionrangeslider"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ meta      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ script    : chr "js/ion.rangeSlider.min.js"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ stylesheet: chr [1:2] "css/ion.rangeSlider.css" "css/ion.rangeSlider.skinShiny.css"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ head      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ attachment: NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ package   : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ all_files : logi TRUE
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "html_dependency"
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 10
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ name      : chr "strftime"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ version   : chr "0.9.2"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ src       :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ href: chr "shared/strftime"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ meta      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ script    : chr "strftime-min.js"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ stylesheet: NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ head      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ attachment: NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ package   : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ all_files : logi TRUE
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "html_dependency"
#>   .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-group shiny-input-container"
#>   .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "control-label"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ for  : chr "my_id-targets_only"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "targets_only"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : Named list()
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "select"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id: chr "my_id-targets_only"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : 'html' chr "<option value=\"TRUE\">TRUE</option>\n<option value=\"FALSE\" selected>FALSE</option>"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "html")= logi TRUE
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "script"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type         : chr "application/json"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-for     : chr "my_id-targets_only"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-nonempty: chr ""
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "{}"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. ..- attr(*, "html_dependencies")=List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 10
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ name      : chr "selectize"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ version   : chr "0.11.2"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ src       :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ href: chr "shared/selectize"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ meta      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ script    : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ stylesheet: chr "css/selectize.bootstrap3.css"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ head      : chr "<!--[if lt IE 9]>\n<script src=\"shared/selectize/js/es5-shim.min.js\"></script>\n<![endif]-->\n<script src=\"s"| __truncated__
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ attachment: NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ package   : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ all_files : logi TRUE
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "html_dependency"
#>   .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-group shiny-input-container"
#>   .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "control-label"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ for  : chr "my_id-outdated"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "outdated"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : Named list()
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "select"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id: chr "my_id-outdated"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : 'html' chr "<option value=\"TRUE\" selected>TRUE</option>\n<option value=\"FALSE\">FALSE</option>"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "html")= logi TRUE
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "script"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type         : chr "application/json"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-for     : chr "my_id-outdated"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-nonempty: chr ""
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "{}"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. ..- attr(*, "html_dependencies")=List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 10
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ name      : chr "selectize"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ version   : chr "0.11.2"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ src       :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ href: chr "shared/selectize"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ meta      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ script    : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ stylesheet: chr "css/selectize.bootstrap3.css"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ head      : chr "<!--[if lt IE 9]>\n<script src=\"shared/selectize/js/es5-shim.min.js\"></script>\n<![endif]-->\n<script src=\"s"| __truncated__
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ attachment: NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ package   : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ all_files : logi TRUE
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "html_dependency"
#>   .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-group shiny-input-container"
#>   .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "control-label"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ for  : chr "my_id-label"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "label"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : Named list()
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "select"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id      : chr "my_id-label"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ multiple: chr "multiple"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : 'html' chr "<option value=\"time\">time</option>\n<option value=\"size\">size</option>\n<option value=\"branches\">branches</option>"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "html")= logi TRUE
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "script"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type    : chr "application/json"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-for: chr "my_id-label"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "{}"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. ..- attr(*, "html_dependencies")=List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 10
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ name      : chr "selectize"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ version   : chr "0.11.2"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ src       :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ href: chr "shared/selectize"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ meta      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ script    : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ stylesheet: chr "css/selectize.bootstrap3.css"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ head      : chr "<!--[if lt IE 9]>\n<script src=\"shared/selectize/js/es5-shim.min.js\"></script>\n<![endif]-->\n<script src=\"s"| __truncated__
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ attachment: NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ package   : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ all_files : logi TRUE
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "html_dependency"
#>   .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "form-group shiny-input-container"
#>   .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "label"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "control-label"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ for  : chr "my_id-level_separation"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "level_separation"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "input"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 13
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class                  : chr "js-range-slider"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id                     : chr "my_id-level_separation"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-min               : chr "0"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-max               : chr "1000"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-from              : chr "150"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-step              : chr "10"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-grid              : chr "true"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-grid-num          : num 10
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-grid-snap         : chr "false"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-prettify-separator: chr ","
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-prettify-enabled  : chr "true"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-keyboard          : chr "true"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-data-type         : chr "number"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. ..- attr(*, "html_dependencies")=List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 10
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ name      : chr "ionrangeslider"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ version   : chr "2.1.6"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ src       :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ href: chr "shared/ionrangeslider"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ meta      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ script    : chr "js/ion.rangeSlider.min.js"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ stylesheet: chr [1:2] "css/ion.rangeSlider.css" "css/ion.rangeSlider.skinShiny.css"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ head      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ attachment: NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ package   : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ all_files : logi TRUE
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "html_dependency"
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 10
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ name      : chr "strftime"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ version   : chr "0.9.2"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ src       :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ href: chr "shared/strftime"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ meta      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ script    : chr "strftime-min.js"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ stylesheet: NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ head      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ attachment: NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ package   : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ all_files : logi TRUE
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "html_dependency"
#>   .. .. .. .. .. .. .. .. .. ..$ : NULL
#>   .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. ..$ : NULL
#>   .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
#>   .. ..- attr(*, "class")= chr "shiny.tag"
#>   ..$ :List of 3
#>   .. ..$ name    : chr "div"
#>   .. ..$ attribs :List of 1
#>   .. .. ..$ class: chr "col-sm-8"
#>   .. ..$ children:List of 1
#>   .. .. ..$ :List of 2
#>   .. .. .. ..$ :List of 3
#>   .. .. .. .. ..$ name    : chr "head"
#>   .. .. .. .. ..$ attribs : Named list()
#>   .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. ..$ name    : chr "style"
#>   .. .. .. .. .. .. ..$ attribs : Named list()
#>   .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. ..$ : 'html' chr ".direct-chat-contacts {\n                 -webkit-transform: translate(100%, 0);\n                 -ms-transfor"| __truncated__
#>   .. .. .. .. .. .. .. .. ..- attr(*, "html")= logi TRUE
#>   .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. ..- attr(*, "htmltools.singleton")= logi TRUE
#>   .. .. .. ..$ :List of 3
#>   .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. ..$ class: chr "col-sm-12"
#>   .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. ..$ class: chr "card card-primary"
#>   .. .. .. .. .. .. ..$ children:List of 3
#>   .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. .. .. ..$ class: chr "card-header"
#>   .. .. .. .. .. .. .. .. ..$ children:List of 2
#>   .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "h3"
#>   .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "card-title"
#>   .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Graph"
#>   .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "card-tools"
#>   .. .. .. .. .. .. .. .. .. .. ..$ children:List of 6
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "button"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ type            : chr "button"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class           : chr "btn btn-tool"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ data-card-widget: chr "collapse"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "i"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "fa fa-minus"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "html_dependencies")=List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 10
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name      : chr "font-awesome"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ version   : chr "5.13.0"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ src       :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ file: chr "www/shared/fontawesome"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ meta      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ script    : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ stylesheet: chr [1:2] "css/all.min.css" "css/v4-shims.min.css"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ head      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attachment: NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ package   : chr "shiny"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ all_files : logi TRUE
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "html_dependency"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "browsable_html")= logi TRUE
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ : NULL
#>   .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. .. .. ..$ attribs :List of 2
#>   .. .. .. .. .. .. .. .. .. ..$ class  : chr "card-body"
#>   .. .. .. .. .. .. .. .. .. ..$ inputID: chr "my_id-graph"
#>   .. .. .. .. .. .. .. .. ..$ children:List of 2
#>   .. .. .. .. .. .. .. .. .. ..$ :List of 4
#>   .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "head"
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : Named list()
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "link"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ rel : chr "stylesheet"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ href: chr "shinycssloaders-assets/spinner.css"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "script"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ src: chr "shinycssloaders-assets/spinner.js"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "htmltools.singleton")= logi TRUE
#>   .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "head"
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : Named list()
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "link"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ rel : chr "stylesheet"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ href: chr "shinycssloaders-assets/css-loaders.css"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "htmltools.singleton")= logi TRUE
#>   .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "head"
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : Named list()
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "style"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs : Named list()
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : 'html' chr "#spinner-4b81c9602724ea14e7f3b60f83da7df3, #spinner-4b81c9602724ea14e7f3b60f83da7df3:before, #spinner-4b81c9602"| __truncated__
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "html")= logi TRUE
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "shiny-spinner-output-container shiny-spinner-hideui "
#>   .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "load-container shiny-spinner-hidden load1"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 2
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "spinner-4b81c9602724ea14e7f3b60f83da7df3"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "loader"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children:List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ : chr "Loading..."
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ : Named list()
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
#>   .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name    : chr "div"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attribs :List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ id   : chr "my_id-graph"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ style: chr "width:100%; height:700px; "
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ class: chr "visNetwork html-widget html-widget-output"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ children: list()
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "html_dependencies")=List of 3
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 10
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name      : chr "htmlwidgets"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ version   : chr "1.5.3"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ src       :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ file: chr "/Users/runner/work/_temp/Library/htmlwidgets/www"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ meta      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ script    : chr "htmlwidgets.js"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ stylesheet: NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ head      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attachment: NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ package   : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ all_files : logi TRUE
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "html_dependency"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 10
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name      : chr "vis"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ version   : chr "4.20.1"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ src       :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ file: chr "/Users/runner/work/_temp/Library/visNetwork/htmlwidgets/lib/vis"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ meta      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ script    : chr "vis.min.js"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ stylesheet: chr "vis.css"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ head      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attachment: NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ package   : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ all_files : logi TRUE
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "html_dependency"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ :List of 10
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ name      : chr "visNetwork-binding"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ version   : chr "2.0.9"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ src       :List of 1
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ file: chr "/Users/runner/work/_temp/Library/visNetwork/htmlwidgets"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ meta      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ script    : chr "visNetwork.js"
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ stylesheet: NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ head      : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ attachment: NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ package   : NULL
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..$ all_files : logi FALSE
#>   .. .. .. .. .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "html_dependency"
#>   .. .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
#>   .. .. .. .. .. .. .. .. .. ..$ : NULL
#>   .. .. .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. .. .. .. ..$ : NULL
#>   .. .. .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. .. ..- attr(*, "class")= chr "shiny.tag"
#>   .. .. .. ..- attr(*, "class")= chr [1:2] "shiny.tag.list" "list"
#>   .. ..- attr(*, "class")= chr "shiny.tag"
#>  - attr(*, "class")= chr "shiny.tag"