Load a saved workspace and seed for debugging.

Load the packages, workspace, and random number generator seed of an errored target attempted with error = "workspace". Remove workspace files with tar_workspaces_destroy() when you are done debugging.

tar_workspace(name, envir = parent.frame(), packages = TRUE, source = TRUE)

Arguments

name
Symbol, name of the target whose workspace to read.

envir
Environment in which to put the objects.

packages
Logical, whether to load the required packages of the target.

source
Logical, whether to run _targets.R to load user-defined global object dependencies into envir. If TRUE, then envir should either be the global environment or inherit from the global environment.

Value

the function returns no value, but it does load the target's required packages, as well as multiple objects into the environment (envir argument) in order to replicate the workspace where the error happened. These objects include the global objects at the time tar_make() was called and the dependency targets. The random number generator seed for the target is also assigned with set.seed()

Details

If you set error = "workspace" in tar_option_set() or tar_target(), then if that target throws an error in tar_make(), it will save its workspace to an RDS file in _targets/workspaces/. Workspaces also get saved for targets supplied to the workspace argument of tar_option_set(), even individual branches. The workspace is a compact reference that allows tar_workspace() to load the target's dependencies and random number generator seed as long as the data objects are still in the data store (usually files in _targets/objects/).

Examples

if (FALSE) {
tmp <- sample(1)
tar_script({
  tar_option_set(error = "workspace")
  tar_pipeline(
    tar_target(x, "loaded"),
    tar_target(y, stop(x))
  )
})
try(tar_make())
exists("x") # Should be FALSE.
tail(.Random.seed)
tar_workspace(y)
exists("x") # Should be TRUE.
print(x) # "loaded"
tail(.Random.seed) # Should be different.
}